adds isolate_decoder
