+++ lib/src/decoder/qr_code_dart_scan_decoder.dart
@@ -1,9 +1,6 @@
 import 'package:camera/camera.dart';
-import 'package:flutter/foundation.dart';
+import 'package:qr_code_dart_scan/src/decoder/isolate_decoder.dart';
 import 'package:zxing_lib/zxing.dart';
-
-import 'decode_event.dart';
-import 'global_functions.dart';
 
 ///
 /// Created by
@@ -30,6 +27,7 @@ class QRCodeDartScanDecoder {
     BarcodeFormat.EAN_13,
   ];
   final List<BarcodeFormat> formats;
+  late IsolateDecoder _isolateDecoder;
 
   QRCodeDartScanDecoder({required this.formats}) {
     for (var format in formats) {
@@ -37,25 +35,19 @@ class QRCodeDartScanDecoder {
         throw Exception('$format format not supported in the moment');
       }
     }
+    _isolateDecoder = IsolateDecoder(formats: formats);
   }
 
   Future<Result?> decodeCameraImage(
     CameraImage image, {
     bool scanInverted = false,
   }) async {
-    final event = DecodeCameraImageEvent(
-      cameraImage: image,
-      formats: formats,
-    );
-    Result? decoded = await compute(
-      decode,
-      event.toMap(),
-    );
+    Result? decoded = await _isolateDecoder.decodeCameraImage(image);
 
     if (scanInverted && decoded == null) {
-      decoded = await compute(
-        decode,
-        event.copyWith(invert: true).toMap(),
+      decoded = await _isolateDecoder.decodeCameraImage(
+        image,
+        insverted: scanInverted,
       );
     }
 
@@ -66,25 +58,10 @@ class QRCodeDartScanDecoder {
     XFile file, {
     bool scanInverted = false,
   }) async {
-    final bytes = await file.readAsBytes();
-
-    final image = await decodeImageFromList(bytes);
-    final event = DecodeImageEvent(
-      image: (await image.toByteData())!.buffer.asUint8List(),
-      width: image.width,
-      height: image.height,
-      formats: formats,
-    );
-    Result? decoded = await compute(
-      decodeImage,
-      event.toMap(),
-    );
+    Result? decoded = await _isolateDecoder.decodeFileImage(file);
 
     if (scanInverted && decoded == null) {
-      decoded = await compute(
-        decodeImage,
-        event.copyWith(invert: true).toMap(),
-      );
+      decoded = await _isolateDecoder.decodeFileImage(file, insverted: true);
     }
 
     return decoded;
